RFC: Add bindings for Rust language
This is just a basic ugly support, not meant to be pushed (at least for now).
There is quite a lot missing, but there is an example which shows that it really
works.  And valgrind reports that all allocations were freed.

The way the code is generated is also not nice, I wish there was more code
actually written in some files and not generated by the generator (as much
hard-coded static strings as possible), maybe similarly to the states.c, I don't
know.

It follows the convention of `-sys` packages:

  https://doc.rust-lang.org/cargo/reference/build-scripts.html#a-sys-packages

But that's about it, I hadn't really had much more time to do anything, I had
only limited free time during the weekend to push this as far as possible and I
got stuck couple of times.  But I think the hardest parts are solved and

Also up for discussion is whether the libnbd crate should be separate since the
higher-level functionality it should provide will not be tightly coupled with
libnbd itself and the releases (especially the numbers) do not need to happen in
sync.

Maybe even the libnbd-sys crate should be in separate repo which has libnbd as a
submodule (in order to use the valuable structuring of API data that is not
installed in the system).

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>